+++ opentelemetry-sdk/src/logs/in_memory_exporter.rs
@@ -15,7 +15,7 @@ use std::sync::{Arc, Mutex};
 /// ```no_run
 ///# use opentelemetry_sdk::logs::{BatchLogProcessor, LoggerProvider};
 ///# use opentelemetry_sdk::runtime;
-///# use opentelemetry_sdk::testing::logs::InMemoryLogExporter;
+///# use opentelemetry_sdk::logs::InMemoryLogExporter;
 ///
 ///# #[tokio::main]
 ///# async fn main() {
@@ -72,7 +72,7 @@ pub struct LogDataWithResource {
 /// # Example
 ///
 /// ```no_run
-///# use opentelemetry_sdk::testing::logs::{InMemoryLogExporter, InMemoryLogExporterBuilder};
+///# use opentelemetry_sdk::logs::{InMemoryLogExporter, InMemoryLogExporterBuilder};
 ///# use opentelemetry_sdk::logs::{BatchLogProcessor, LoggerProvider};
 ///# use opentelemetry_sdk::runtime;
 ///
@@ -139,7 +139,7 @@ impl InMemoryLogExporter {
     /// # Example
     ///
     /// ```
-    /// use opentelemetry_sdk::testing::logs::{InMemoryLogExporter, InMemoryLogExporterBuilder};
+    /// use opentelemetry_sdk::logs::{InMemoryLogExporter, InMemoryLogExporterBuilder};
     ///
     /// let exporter = InMemoryLogExporterBuilder::default().build();
     /// let emitted_logs = exporter.get_emitted_logs().unwrap();
@@ -164,7 +164,7 @@ impl InMemoryLogExporter {
     /// # Example
     ///
     /// ```
-    /// use opentelemetry_sdk::testing::logs::{InMemoryLogExporter, InMemoryLogExporterBuilder};
+    /// use opentelemetry_sdk::logs::{InMemoryLogExporter, InMemoryLogExporterBuilder};
     ///
     /// let exporter = InMemoryLogExporterBuilder::default().build();
     /// exporter.reset();