+++ opentelemetry-sdk/src/trace/provider.rs
@@ -66,7 +66,7 @@ use crate::trace::{
     BatchSpanProcessor, Config, RandomIdGenerator, Sampler, SimpleSpanProcessor, SpanLimits, Tracer,
 };
 use crate::Resource;
-use crate::{export::trace::SpanExporter, trace::SpanProcessor};
+use crate::{trace::SpanExporter, trace::SpanProcessor};
 use opentelemetry::trace::TraceError;
 use opentelemetry::{otel_debug, trace::TraceResult};
 use opentelemetry::{otel_info, InstrumentationScope};
@@ -443,11 +443,11 @@ impl Builder {
 
 #[cfg(test)]
 mod tests {
-    use crate::export::trace::SpanData;
     use crate::resource::{
         SERVICE_NAME, TELEMETRY_SDK_LANGUAGE, TELEMETRY_SDK_NAME, TELEMETRY_SDK_VERSION,
     };
     use crate::trace::provider::TracerProviderInner;
+    use crate::trace::SpanData;
     use crate::trace::{Config, Span, SpanProcessor};
     use crate::Resource;
     use opentelemetry::trace::{TraceError, TraceResult, Tracer, TracerProvider};
@@ -552,7 +552,7 @@ mod tests {
                 provider
                     .config()
                     .resource
-                    .get(Key::from_static_str(resource_key))
+                    .get(&Key::from_static_str(resource_key))
                     .map(|v| v.to_string()),
                 expect.map(|s| s.to_string())
             );
@@ -562,15 +562,18 @@ mod tests {
                 provider
                     .config()
                     .resource
-                    .get(TELEMETRY_SDK_LANGUAGE.into()),
+                    .get(&TELEMETRY_SDK_LANGUAGE.into()),
                 Some(Value::from("rust"))
             );
             assert_eq!(
-                provider.config().resource.get(TELEMETRY_SDK_NAME.into()),
+                provider.config().resource.get(&TELEMETRY_SDK_NAME.into()),
                 Some(Value::from("opentelemetry"))
             );
             assert_eq!(
-                provider.config().resource.get(TELEMETRY_SDK_VERSION.into()),
+                provider
+                    .config()
+                    .resource
+                    .get(&TELEMETRY_SDK_VERSION.into()),
                 Some(Value::from(env!("CARGO_PKG_VERSION")))
             );
         };